refactor: splitted traits into multiple files
