+++ chain/jsonrpc/src/api/gas_price.rs
@@ -4,7 +4,7 @@ use serde_json::Value;
 use near_client_primitives::types::GetGasPriceError;
 use near_jsonrpc_primitives::errors::RpcParseError;
 use near_jsonrpc_primitives::types::gas_price::{RpcGasPriceError, RpcGasPriceRequest};
-use near_jsonrpc_traits::ParamsTrait;
+use near_jsonrpc_traits::params::ParamsTrait;
 
 use super::{Params, RpcFrom, RpcRequest};
