+++ chain/jsonrpc/src/api/sandbox.rs
@@ -6,7 +6,7 @@ use near_jsonrpc_primitives::types::sandbox::{
     RpcSandboxFastForwardError, RpcSandboxFastForwardRequest, RpcSandboxPatchStateError,
     RpcSandboxPatchStateRequest,
 };
-use near_jsonrpc_traits::ParamsTrait;
+use near_jsonrpc_traits::params::ParamsTrait;
 
 use super::{Params, RpcFrom, RpcRequest};
